requirements.txt
# FATES-Emulator Requirements
# Install with: pip install -r requirements.txt

# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
xgboost>=1.5.0
scikit-learn>=1.0.0
flaml>=1.0.0  # AutoML for hyperparameter optimization
lightgbm>=3.3.0  # Additional estimator for FLAML

# Interpretability
shap>=0.40.0

# Sensitivity analysis
SALib>=1.4.0

# Data handling
xarray>=0.19.0
netCDF4>=1.5.7

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration and utilities
pyyaml>=5.4.0
tqdm>=4.62.0

# Jupyter (for examples)
jupyter>=1.0.0
ipykernel>=6.4.0

# Development and testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.9b0
flake8>=3.9.0

